refactor: optimize comments, delete useless files, and move test files
